+++ docker/const.sh
@@ -1,6 +1,0 @@
-#! /bin/bash
-
-export CONTAINER_NAME="sola_mpd"
-export IMAGE_PREFIX="sola_mpd_prebuilt"
-export IMAGE_TAG="latest"
-export IMAGE_NAME="${IMAGE_PREFIX}:${IMAGE_TAG}"